refactor chart-builder.tsx
