+++ apps/shared/src/chart-builder.tsx
@@ -22,7 +22,18 @@ import {
 	YAxis,
 } from 'recharts';
 
+import {
+	DATA_LABEL_MARGIN_TOP,
+	DATA_LABEL_X_AXIS_FOOTROOM,
+	DataLabelsLayer,
+	getDataLabelSetup,
+	PIE_LABELLED_OUTER_RADIUS,
+	PieDataLabelsLayer,
+	shouldReserveDataLabelHeadroom,
+	shouldReserveStackTotalFootroom,
+} from './chart-data-labels';
 import { collectAxisValues, collectStackedAxisValues, resolveYAxisDomain } from './chart-domain';
+import { formatCompactNumber, toFiniteNumber } from './chart-values';
 import { type DateFormatSettings, formatDateValue, isIsoDateLike } from './date';
 import * as displayChart from './tools/display-chart';
 
@@ -30,30 +41,6 @@ export const DEFAULT_COLORS = ['#104e64', '#f54900', '#009689', '#ffb900', '#fe9
 
 const AXIS_TICK = { fontSize: 12 };
 const CATEGORY_XAXIS_HEIGHT = 56;
-const DATA_LABEL_PROPS = {
-	fill: 'var(--foreground, #111827)',
-	fontSize: 11,
-	fontFamily: 'system-ui, sans-serif',
-};
-const DATA_LABEL_MARGIN_TOP = 24;
-const DATA_LABEL_X_AXIS_FOOTROOM = 24;
-const DATA_LABEL_HEADROOM_RATIO = 0.9;
-
-const DATA_LABEL_FONT_SIZE = 11;
-/** Approximate glyph width as a fraction of the font size; used to size collision boxes without a DOM. */
-const DATA_LABEL_CHAR_WIDTH_RATIO = 0.6;
-/** Vertical clearance between the anchor point/line (or bar edge) and the nearest edge of a label. */
-const DATA_LABEL_ANCHOR_GAP = 8;
-/** Padding added around each label's collision box so neighbours keep a little breathing room. */
-const DATA_LABEL_BOX_PADDING = 2;
-/** Series kinds whose points we place labels for; stacked/pie use their own paths. */
-const LABELLED_SERIES_KINDS = new Set(['Bar', 'Line', 'Area']);
-const MAX_SERIES_DATA_LABELS = 12;
-const PIE_LABELLED_OUTER_RADIUS = '65%';
-const PIE_DATA_LABEL_GAP = 10;
-
-/** Theme-aware background used to draw the thin gaps between pie/donut slices. */
-const DEFAULT_BACKGROUND = 'var(--background, #ffffff)';
 
 /** Beyond this many slices, pie/donut charts bucket the smallest into a single "Other" slice. */
 const MAX_PIE_SLICES = 10;
@@ -68,7 +55,7 @@ const STACK_SEPARATOR_WIDTH = 1;
  * in the browser; the concrete fallback covers the backend PNG/HTML export where the backend
  * passes an explicit `backgroundColor` and CSS vars do not resolve.
  */
-const DEFAULT_BACKGROUND_COLOR = 'var(--background, #ffffff)';
+const DEFAULT_BACKGROUND = 'var(--background, #ffffff)';
 
 /**
  * Reserved width for the Y axis band. Smaller than the Recharts default (60)
@@ -83,20 +70,6 @@ export function labelize(key: unknown, dateFormat?: DateFormatSettings | null): 
 		return formatDateValue(str, dateFormat);
 	}
 	return str.replace(/_/g, ' ').replace(/\b\w/g, (c) => c.toUpperCase());
-}
-
-export function formatCompactNumber(value: number): string {
-	const abs = Math.abs(value);
-	if (abs >= 1_000_000_000) {
-		return `${(value / 1_000_000_000).toFixed(1).replace(/\.0$/, '')}B`;
-	}
-	if (abs >= 1_000_000) {
-		return `${(value / 1_000_000).toFixed(1).replace(/\.0$/, '')}M`;
-	}
-	if (abs >= 10_000) {
-		return `${(value / 1_000).toFixed(1).replace(/\.0$/, '')}K`;
-	}
-	return value.toLocaleString();
 }
 
 /**
@@ -254,11 +227,6 @@ function parseDateMs(value: unknown): number | null {
 	return Number.isNaN(ms) ? null : ms;
 }
 
-export function formatDataLabel(value: unknown): string {
-	const number = toFiniteNumber(value);
-	return number == null ? '' : formatCompactNumber(number);
-}
-
 export function defaultColorFor(_key: string, index: number): string {
 	return DEFAULT_COLORS[index % DEFAULT_COLORS.length];
 }
@@ -414,96 +382,6 @@ function buildChartMargin(props: BuildChartProps, showTitle: boolean) {
 		return props.margin;
 	}
 	return { ...props.margin, top: (props.margin?.top ?? 0) + titleTop + labelsTop };
-}
-
-export function shouldReserveDataLabelHeadroom(props: BuildChartProps): boolean {
-	if (props.showDataLabels !== true || !isCartesianLabelChart(props.chartType)) {
-		return false;
-	}
-	// Stacked charts label the running total at the very top of the stack, which can sit flush against
-	// the axis top, so always keep room for that label rather than relying on the axis-headroom ratio.
-	if (displayChart.isStackedChartType(props.chartType)) {
-		return true;
-	}
-	const maxValue = getMaxPlottedValue(props);
-	if (maxValue == null || maxValue <= 0) {
-		return false;
-	}
-	return maxValue >= niceAxisMax(maxValue) * DATA_LABEL_HEADROOM_RATIO;
-}
-
-function shouldReserveStackTotalFootroom(props: BuildChartProps): boolean {
-	if (props.showDataLabels !== true || !displayChart.isStackedChartType(props.chartType)) {
-		return false;
-	}
-	if (displayChart.isPercentStackedChartType(props.chartType)) {
-		return false;
-	}
-	return props.data.some((row) => {
-		const total = sumStackValue(row, props.series);
-		return total != null && total < 0;
-	});
-}
-
-function isCartesianLabelChart(chartType: displayChart.ChartType): boolean {
-	return (
-		chartType === 'bar' ||
-		chartType === 'line' ||
-		chartType === 'area' ||
-		displayChart.isStackedChartType(chartType) ||
-		chartType === 'mixed'
-	);
-}
-
-function getMaxPlottedValue(props: BuildChartProps): number | null {
-	const isStacked = displayChart.isStackedChartType(props.chartType);
-	return isStacked ? getMaxStackTotal(props.data, props.series) : getMaxSeriesValue(props.data, props.series);
-}
-
-function getMaxSeriesValue(data: Record<string, unknown>[], series: displayChart.SeriesConfig[]): number | null {
-	let max: number | null = null;
-	for (const row of data) {
-		for (const item of series) {
-			const value = toFiniteNumber(row[item.data_key]);
-			if (value != null && (max == null || value > max)) {
-				max = value;
-			}
-		}
-	}
-	return max;
-}
-
-function getMaxStackTotal(data: Record<string, unknown>[], series: displayChart.SeriesConfig[]): number | null {
-	let max: number | null = null;
-	for (const row of data) {
-		let positive = 0;
-		for (const item of series) {
-			if (item.is_total) {
-				continue;
-			}
-			const value = toFiniteNumber(row[item.data_key]);
-			if (value != null && value > 0) {
-				positive += value;
-			}
-		}
-		if (positive > 0 && (max == null || positive > max)) {
-			max = positive;
-		}
-	}
-	return max;
-}
-
-export function niceAxisMax(dataMax: number, tickCount = 5): number {
-	if (dataMax <= 0) {
-		return 0;
-	}
-	const roughStep = dataMax / (tickCount - 1);
-	const magnitude = 10 ** Math.floor(Math.log10(roughStep));
-	const normalized = roughStep / magnitude;
-	const niceNormalized =
-		normalized <= 1 ? 1 : normalized <= 2 ? 2 : normalized <= 2.5 ? 2.5 : normalized <= 5 ? 5 : 10;
-	const niceStep = niceNormalized * magnitude;
-	return niceStep * Math.ceil(dataMax / niceStep);
 }
 
 function buildKpiCard(props: ResolvedProps) {
@@ -676,7 +554,7 @@ function buildBarChart(props: ResolvedProps) {
 	const axisValues = isStacked ? collectStackedAxisValues(data, dataKeys) : collectAxisValues(data, dataKeys);
 	const { renderedSeries, stackTotalLayer } = getDataLabelSetup(props, isStacked);
 	const seriesKeys = renderedSeries.map((s) => s.data_key);
-	const separatorColor = props.backgroundColor ?? DEFAULT_BACKGROUND_COLOR;
+	const separatorColor = props.backgroundColor ?? DEFAULT_BACKGROUND;
 	const labelFootroom = shouldReserveStackTotalFootroom(props) ? DATA_LABEL_X_AXIS_FOOTROOM : 0;
 
 	return (
@@ -1199,491 +1077,4 @@ function renderChartTitle(title: string) {
 			)}
 		/>
 	);
-}
-
-function isLocalMaximum(values: (number | null)[], index: number): boolean {
-	const value = values[index];
-	if (value == null) {
-		return false;
-	}
-	const left = values[index - 1] ?? null;
-	const right = values[index + 1] ?? null;
-	return (left == null || value > left) && (right == null || value > right);
-}
-
-function isLocalMinimum(values: (number | null)[], index: number): boolean {
-	const value = values[index];
-	if (value == null) {
-		return false;
-	}
-	const left = values[index - 1] ?? null;
-	const right = values[index + 1] ?? null;
-	return (left == null || value < left) && (right == null || value < right);
-}
-
-/** A point is a local extremum when its value turns direction — a peak or a trough versus neighbours. */
-function isLocalExtremum(values: (number | null)[], index: number): boolean {
-	return isLocalMaximum(values, index) || isLocalMinimum(values, index);
-}
-
-interface PlotRect {
-	top?: number;
-	left?: number;
-	width?: number;
-	height?: number;
-}
-
-interface GraphicalPoint {
-	x?: number;
-	y?: number;
-	width?: number;
-	height?: number;
-	value?: unknown;
-}
-
-interface PieSector {
-	cx?: number;
-	cy?: number;
-	innerRadius?: number;
-	outerRadius?: number;
-	midAngle?: number;
-	percent?: number;
-	value?: unknown;
-}
-
-interface GraphicalItem {
-	item?: { type?: { displayName?: string } };
-	props?: { points?: GraphicalPoint[]; data?: GraphicalPoint[]; sectors?: PieSector[] };
-}
-
-interface DataLabelsLayerProps {
-	formattedGraphicalItems?: GraphicalItem[];
-	offset?: PlotRect;
-	width?: number;
-	height?: number;
-}
-
-interface LabelCandidate {
-	cx: number;
-	baselineY: number;
-	box: LabelBox;
-	text: string;
-	rank: number[];
-	textAnchor: 'start' | 'middle' | 'end';
-}
-
-interface IndexedLabelCandidate {
-	candidate: LabelCandidate;
-	index: number;
-	value: number;
-}
-
-interface LabelBox {
-	left: number;
-	right: number;
-	top: number;
-	bottom: number;
-}
-
-/** Coordinates cartesian labels across series so conflicts can be dropped without moving labels. */
-function DataLabelsLayer({ formattedGraphicalItems, offset }: DataLabelsLayerProps) {
-	const candidates = collectLabelCandidates(formattedGraphicalItems ?? [], offset ?? {});
-	return renderLabels(resolveLabelOverlaps(candidates));
-}
-
-function renderStackTotalLabelsLayer(data: Record<string, unknown>[], series: displayChart.SeriesConfig[]) {
-	return function StackTotalLabelsLayer({ formattedGraphicalItems, offset }: DataLabelsLayerProps) {
-		const items = (formattedGraphicalItems ?? []).filter((entry) =>
-			LABELLED_SERIES_KINDS.has(entry.item?.type?.displayName ?? ''),
-		);
-		const kind = items[0]?.item?.type?.displayName;
-		if (!kind || items.length === 0) {
-			return null;
-		}
-		const candidates = stackTotalCandidates(items, data, series, kind === 'Bar', offset ?? {});
-		return renderLabels(resolveLabelOverlaps(candidates));
-	};
-}
-
-function PieDataLabelsLayer({ formattedGraphicalItems, width, height }: DataLabelsLayerProps) {
-	const pie = (formattedGraphicalItems ?? []).find((entry) => entry.item?.type?.displayName === 'Pie');
-	const candidates = collectPieLabelCandidates(pie?.props?.sectors ?? [], width, height);
-	return renderLabels(resolveLabelOverlaps(candidates));
-}
-
-function renderLabels(labels: LabelCandidate[]) {
-	if (labels.length === 0) {
-		return null;
-	}
-	return (
-		<g className='recharts-data-labels'>
-			{labels.map((label, index) => (
-				<text
-					key={index}
-					x={label.cx}
-					y={label.baselineY}
-					textAnchor={label.textAnchor}
-					dominantBaseline='alphabetic'
-					{...DATA_LABEL_PROPS}
-				>
-					{label.text}
-				</text>
-			))}
-		</g>
-	);
-}
-
-function stackTotalCandidates(
-	items: GraphicalItem[],
-	data: Record<string, unknown>[],
-	series: displayChart.SeriesConfig[],
-	isBar: boolean,
-	plot: PlotRect,
-): LabelCandidate[] {
-	const totals = data.map((row) => sumStackValue(row, series));
-
-	return totals.flatMap((total, dataIndex) => {
-		if (total == null || total === 0) {
-			return [];
-		}
-		const isPositive = total > 0;
-		const anchor = stackTotalAnchor(items, dataIndex, isBar, isPositive);
-		if (!anchor) {
-			return [];
-		}
-		const baselineY = isPositive
-			? anchor.anchorY - DATA_LABEL_ANCHOR_GAP
-			: anchor.anchorY + DATA_LABEL_ANCHOR_GAP + DATA_LABEL_FONT_SIZE;
-		const candidate = buildLabelCandidate(
-			anchor.cx,
-			baselineY,
-			formatDataLabel(total),
-			cartesianLabelRank(isLocalExtremum(totals, dataIndex), 0, anchor.cx),
-			plot,
-			isPositive ? 0 : DATA_LABEL_X_AXIS_FOOTROOM,
-		);
-		return candidate ? [candidate] : [];
-	});
-}
-
-function stackTotalAnchor(
-	items: GraphicalItem[],
-	dataIndex: number,
-	isBar: boolean,
-	isPositive: boolean,
-): { cx: number; anchorY: number } | null {
-	let extreme: { cx: number; anchorY: number } | null = null;
-	items.forEach((item) => {
-		const points = isBar ? item.props?.data : item.props?.points;
-		const point = points?.[dataIndex];
-		const value = pointStackValue(point);
-		if (value == null || (isPositive ? value <= 0 : value >= 0)) {
-			return;
-		}
-		const x = toFiniteNumber(point?.x);
-		const y = toFiniteNumber(point?.y);
-		if (point == null || x == null || y == null) {
-			return;
-		}
-		const width = isBar ? (toFiniteNumber(point.width) ?? 0) : 0;
-		const otherY = y + (isBar ? (toFiniteNumber(point.height) ?? 0) : 0);
-		const anchor = {
-			cx: x + width / 2,
-			anchorY: isPositive ? Math.min(y, otherY) : Math.max(y, otherY),
-		};
-		if (extreme == null || (isPositive ? anchor.anchorY < extreme.anchorY : anchor.anchorY > extreme.anchorY)) {
-			extreme = anchor;
-		}
-	});
-	return extreme;
-}
-
-function pointStackValue(point: GraphicalPoint | undefined): number | null {
-	if (!Array.isArray(point?.value) || point.value.length < 2) {
-		return toFiniteNumber(point?.value);
-	}
-	const start = toFiniteNumber(point.value[0]);
-	const end = toFiniteNumber(point.value[point.value.length - 1]);
-	return start == null || end == null ? null : end - start;
-}
-
-function collectPieLabelCandidates(
-	sectors: PieSector[],
-	width: number | undefined,
-	height: number | undefined,
-): LabelCandidate[] {
-	if (width == null || height == null) {
-		return [];
-	}
-	return sectors.flatMap((sector, sliceIndex) => {
-		const cx = toFiniteNumber(sector.cx);
-		const cy = toFiniteNumber(sector.cy);
-		const outerRadius = toFiniteNumber(sector.outerRadius);
-		const midAngle = toFiniteNumber(sector.midAngle);
-		const text = formatDataLabel(sector.value);
-		if (cx == null || cy == null || outerRadius == null || midAngle == null || !text) {
-			return [];
-		}
-		const radius = outerRadius + PIE_DATA_LABEL_GAP;
-		const angle = (-midAngle * Math.PI) / 180;
-		const labelX = cx + radius * Math.cos(angle);
-		const baselineY = cy + radius * Math.sin(angle);
-		const textAnchor = labelX >= cx ? 'start' : 'end';
-		const halfWidth = (text.length * DATA_LABEL_FONT_SIZE * DATA_LABEL_CHAR_WIDTH_RATIO) / 2;
-		const box = labelBox(labelX, baselineY, halfWidth, textAnchor);
-		if (box.left < 0 || box.right > width || box.top < 0 || box.bottom > height) {
-			return [];
-		}
-		return [
-			{
-				cx: labelX,
-				baselineY,
-				box,
-				text,
-				rank: [sector.percent ?? 0, -sliceIndex],
-				textAnchor,
-			},
-		];
-	});
-}
-
-function collectLabelCandidates(items: GraphicalItem[], plot: PlotRect): LabelCandidate[] {
-	return items.flatMap((entry, seriesIndex) => {
-		const kind = entry.item?.type?.displayName;
-		if (!kind || !LABELLED_SERIES_KINDS.has(kind)) {
-			return [];
-		}
-		const points = kind === 'Bar' ? entry.props?.data : entry.props?.points;
-		if (!points || points.length === 0) {
-			return [];
-		}
-		return seriesCandidates(points, points.map(pointNumericValue), kind === 'Bar', seriesIndex, plot);
-	});
-}
-
-function seriesCandidates(
-	points: GraphicalPoint[],
-	values: (number | null)[],
-	isBar: boolean,
-	seriesIndex: number,
-	plot: PlotRect,
-): LabelCandidate[] {
-	const candidates = points.flatMap((point, index): IndexedLabelCandidate[] => {
-		const value = values[index];
-		if (value == null) {
-			return [];
-		}
-		const text = formatDataLabel(value);
-		if (!text) {
-			return [];
-		}
-		const anchor = labelAnchor(point, value, isBar);
-		if (anchor == null) {
-			return [];
-		}
-		const candidate = buildLabelCandidate(
-			anchor.cx,
-			anchor.baselineY,
-			text,
-			cartesianLabelRank(isLocalExtremum(values, index), seriesIndex, anchor.cx),
-			plot,
-		);
-		return candidate ? [{ candidate, index, value }] : [];
-	});
-	if (isBar || candidates.length <= MAX_SERIES_DATA_LABELS) {
-		return candidates.map((entry) => entry.candidate);
-	}
-	return selectSeriesLabelCandidates(candidates, values).map((entry) => entry.candidate);
-}
-
-function selectSeriesLabelCandidates(
-	candidates: IndexedLabelCandidate[],
-	values: (number | null)[],
-): IndexedLabelCandidate[] {
-	const selectedIndices = new Set<number>();
-	const minimum = candidates.reduce((best, current) => (current.value < best.value ? current : best));
-	const maximum = candidates.reduce((best, current) => (current.value > best.value ? current : best));
-	selectedIndices.add(minimum.index);
-	selectedIndices.add(maximum.index);
-
-	const remainingSlots = () => MAX_SERIES_DATA_LABELS - selectedIndices.size;
-	const extrema = candidates.filter(
-		(entry) => !selectedIndices.has(entry.index) && isLocalExtremum(values, entry.index),
-	);
-	const selectedExtrema =
-		extrema.length <= remainingSlots() ? extrema : sampleLabelCandidatesEvenly(extrema, remainingSlots());
-	selectedExtrema.forEach((entry) => selectedIndices.add(entry.index));
-
-	const remaining = candidates.filter((entry) => !selectedIndices.has(entry.index));
-	sampleLabelCandidatesEvenly(remaining, remainingSlots()).forEach((entry) => selectedIndices.add(entry.index));
-	return candidates.filter((entry) => selectedIndices.has(entry.index));
-}
-
-function sampleLabelCandidatesEvenly(candidates: IndexedLabelCandidate[], count: number): IndexedLabelCandidate[] {
-	if (count <= 0) {
-		return [];
-	}
-	if (candidates.length <= count) {
-		return candidates;
-	}
-
-	const ordered = [...candidates].sort((a, b) => a.candidate.cx - b.candidate.cx || a.index - b.index);
-	const firstX = ordered[0].candidate.cx;
-	const lastX = ordered[ordered.length - 1].candidate.cx;
-	const selected = new Set<number>();
-
-	for (let position = 0; position < count; position += 1) {
-		const fraction = count === 1 ? 0.5 : position / (count - 1);
-		const targetX = firstX + (lastX - firstX) * fraction;
-		const closest = ordered
-			.filter((entry) => !selected.has(entry.index))
-			.reduce((best, current) => {
-				const currentDistance = Math.abs(current.candidate.cx - targetX);
-				const bestDistance = Math.abs(best.candidate.cx - targetX);
-				return currentDistance < bestDistance ||
-					(currentDistance === bestDistance && current.index < best.index)
-					? current
-					: best;
-			});
-		selected.add(closest.index);
-	}
-
-	return ordered.filter((entry) => selected.has(entry.index));
-}
-
-function buildLabelCandidate(
-	cx: number,
-	baselineY: number,
-	text: string,
-	rank: number[],
-	plot: PlotRect,
-	bottomAllowance = 0,
-): LabelCandidate | null {
-	const halfWidth = (text.length * DATA_LABEL_FONT_SIZE * DATA_LABEL_CHAR_WIDTH_RATIO) / 2;
-	const box = labelBox(cx, baselineY, halfWidth);
-	if (!boxFitsPlot(box, plot, bottomAllowance)) {
-		return null;
-	}
-	return { cx, baselineY, box, text, rank, textAnchor: 'middle' };
-}
-
-/** Area/line points carry `value` as a `[baseLine, value]` range; bars carry a scalar. Unwrap both. */
-function pointNumericValue(point: GraphicalPoint): number | null {
-	const raw = Array.isArray(point.value) ? point.value[point.value.length - 1] : point.value;
-	return toFiniteNumber(raw);
-}
-
-/** Natural label position: one gap above the point/bar-top (below the bar for negative bars). */
-function labelAnchor(point: GraphicalPoint, value: number, isBar: boolean): { cx: number; baselineY: number } | null {
-	const x = toFiniteNumber(point.x);
-	const y = toFiniteNumber(point.y);
-	if (x == null || y == null) {
-		return null;
-	}
-	if (!isBar) {
-		return { cx: x, baselineY: y - DATA_LABEL_ANCHOR_GAP };
-	}
-	const width = toFiniteNumber(point.width) ?? 0;
-	const height = toFiniteNumber(point.height) ?? 0;
-	return value >= 0
-		? { cx: x + width / 2, baselineY: y - DATA_LABEL_ANCHOR_GAP }
-		: { cx: x + width / 2, baselineY: y + height + DATA_LABEL_ANCHOR_GAP + DATA_LABEL_FONT_SIZE };
-}
-
-/**
- * Keeps labels at their natural position and drops conflicts. If no two boxes intersect, every label
- * renders as-is. Otherwise a greedy pass in priority order keeps the winner of each collision and
- * drops the rest — no nudging, so placement stays predictable.
- */
-function resolveLabelOverlaps(candidates: LabelCandidate[]): LabelCandidate[] {
-	const ordered = [...candidates].sort(byLabelRank);
-	const kept = keepNonOverlapping(ordered);
-	return kept.length === candidates.length ? candidates : kept;
-}
-
-function keepNonOverlapping(ordered: LabelCandidate[]): LabelCandidate[] {
-	const kept: LabelCandidate[] = [];
-	for (const candidate of ordered) {
-		if (!kept.some((other) => boxesOverlap(other.box, candidate.box))) {
-			kept.push(candidate);
-		}
-	}
-	return kept;
-}
-
-function byLabelRank(a: LabelCandidate, b: LabelCandidate): number {
-	for (let index = 0; index < Math.max(a.rank.length, b.rank.length); index += 1) {
-		const difference = (b.rank[index] ?? 0) - (a.rank[index] ?? 0);
-		if (difference !== 0) {
-			return difference;
-		}
-	}
-	return 0;
-}
-
-function cartesianLabelRank(isExtremum: boolean, seriesIndex: number, x: number): number[] {
-	return [isExtremum ? 1 : 0, -seriesIndex, -x];
-}
-
-function boxFitsPlot(box: LabelBox, plot: PlotRect, bottomAllowance = 0): boolean {
-	const minLeft = plot.left ?? Number.NEGATIVE_INFINITY;
-	const maxRight = plot.left != null && plot.width != null ? plot.left + plot.width : Number.POSITIVE_INFINITY;
-	const minTop = plot.top != null ? Math.max(0, plot.top - DATA_LABEL_MARGIN_TOP) : Number.NEGATIVE_INFINITY;
-	const maxBottom =
-		plot.top != null && plot.height != null ? plot.top + plot.height + bottomAllowance : Number.POSITIVE_INFINITY;
-	return box.left >= minLeft && box.right <= maxRight && box.top >= minTop && box.bottom <= maxBottom;
-}
-
-function labelBox(
-	cx: number,
-	baselineY: number,
-	halfWidth: number,
-	textAnchor: LabelCandidate['textAnchor'] = 'middle',
-): LabelBox {
-	const left = textAnchor === 'start' ? cx : textAnchor === 'end' ? cx - halfWidth * 2 : cx - halfWidth;
-	const right = textAnchor === 'start' ? cx + halfWidth * 2 : textAnchor === 'end' ? cx : cx + halfWidth;
-	return {
-		left: left - DATA_LABEL_BOX_PADDING,
-		right: right + DATA_LABEL_BOX_PADDING,
-		top: baselineY - DATA_LABEL_FONT_SIZE - DATA_LABEL_BOX_PADDING,
-		bottom: baselineY + DATA_LABEL_BOX_PADDING,
-	};
-}
-
-function boxesOverlap(a: LabelBox, b: LabelBox): boolean {
-	return a.left < b.right && a.right > b.left && a.top < b.bottom && a.bottom > b.top;
-}
-
-function getDataLabelSetup(props: ResolvedProps, isStacked: boolean) {
-	const renderedSeries = getRenderedSeries(isStacked, props.series);
-	const stackTotalLayer =
-		props.showDataLabels && isStacked && renderedSeries.length > 0
-			? renderStackTotalLabelsLayer(props.data, props.series)
-			: undefined;
-	return { renderedSeries, stackTotalLayer };
-}
-
-function getRenderedSeries(isStacked: boolean, series: displayChart.SeriesConfig[]): displayChart.SeriesConfig[] {
-	return isStacked ? series.filter((item) => !item.is_total) : series;
-}
-
-function sumStackValue(row: Record<string, unknown> | undefined, series: displayChart.SeriesConfig[]): number | null {
-	if (!row) {
-		return null;
-	}
-
-	const values = series.filter((s) => !s.is_total).map((s) => toFiniteNumber(row[s.data_key]));
-	const numericValues = values.filter((value): value is number => value != null);
-	return numericValues.length > 0 ? numericValues.reduce((sum, value) => sum + value, 0) : null;
-}
-
-function toFiniteNumber(value: unknown): number | null {
-	if (typeof value === 'number') {
-		return Number.isFinite(value) ? value : null;
-	}
-	if (typeof value === 'string' && value.trim() !== '') {
-		const parsed = Number(value);
-		return Number.isFinite(parsed) ? parsed : null;
-	}
-	return null;
 }
