+++ apps/shared/src/chart-values.ts
@@ -1,0 +1,24 @@
+export function formatCompactNumber(value: number): string {
+	const abs = Math.abs(value);
+	if (abs >= 1_000_000_000) {
+		return `${(value / 1_000_000_000).toFixed(1).replace(/\.0$/, '')}B`;
+	}
+	if (abs >= 1_000_000) {
+		return `${(value / 1_000_000).toFixed(1).replace(/\.0$/, '')}M`;
+	}
+	if (abs >= 10_000) {
+		return `${(value / 1_000).toFixed(1).replace(/\.0$/, '')}K`;
+	}
+	return value.toLocaleString();
+}
+
+export function toFiniteNumber(value: unknown): number | null {
+	if (typeof value === 'number') {
+		return Number.isFinite(value) ? value : null;
+	}
+	if (typeof value === 'string' && value.trim() !== '') {
+		const parsed = Number(value);
+		return Number.isFinite(parsed) ? parsed : null;
+	}
+	return null;
+}
