+++ apps/shared/src/index.ts
@@ -1,7 +1,9 @@
 export * from './chart-block';
 export * from './chart-builder';
+export * from './chart-data-labels';
 export * from './chart-domain';
 export * from './chart-plugin';
+export * from './chart-values';
 export * from './citation';
 export * from './date';
 export * from './execute-sql-parts';
